Move post-review into a new top-level "rbtools" package
Since the dawn of time, post-review has lived in the Review Board tree, despite being a client script and not necessarily being needed on the server.

This moves post-review into a new rbtools package, which users will be able to easy_install (easy_install -U RBTools). The simplejson dependency will be fetched, and users won't have to go hunting in SVN for a script.

Over time, I'd like to add new scripts into here (such as one to apply a diff from a review request), and further down the road I'd like to create an "rbapi" (or something) package that contains a Python API for talking to a Review Board server, which rbtools (and post-review) will then use. For now, the goal is to just get post-review moved over.

Reviewed at http://reviews.review-board.org/r/788/


git-svn-id: https://reviewboard.googlecode.com/svn/trunk/rbtools@1861 5efc13c4-1f27-0410-8691-ff2d1f55687e