Don't unconditionally publish the draft when setting a field.
When setting a field (such as target reviewers) from post-review, we were
saving the draft, always. This is bad enough, but we were doing this before
even updating the diff. This would result is a published draft and an extra
change description entry. We no longer save the draft, unless the user
uses --publish.

Fixes bug #811