Support creating diffs with multiple files when using --revision-range
When creating an SVN diff for a remote repository using --revision-range and
passing in specific filenames, post-review would only get the first file as the
diff. This is because it was using the arg as the base path. This change makes
it treat the arg as a file list if it's more than one item long.

Fixes bug 1162