Don't produce broken diff files with files that don't end in a newline.
Some people have source files that don't end in a newline. In this case, it's
possible for us to create a diff file where the end of one file is joined on the
same line as the header for the next file. This change adds an extra check to
_do_diff to make sure that we don't make broken diffs when people have broken
files.

Fixes bug 1340.