Fix post-review error reporting when using --diff-only.
Running post-review --diff-only may throw exceptions when encountering
errors, instead of gracefully handlin them. A simple case of this is to
run with --diff-only without having a post-review login cookie, and instead
of a login prompt, post-review will crash with a KeyError.

This change fixes this reporting by not special-casing the --diff-only
parameter in the error reporting. We were only handling errors if that
option was not set.

Patch by Eric Huss.
Reviewed at http://reviews.reviewboard.org/r/1185/