Filter out usernames from Mercurial repository paths in .hgrc
We filtered out the username from a path when running 'hg svn info', but
not when reading from .hgrc. We now apply the same filtering logic. This
allows post-review to match a repository path with a username in it against
a repository path on a Review Board server.

Patch by Tom Saeger.
Reviewed at http://reviews.reviewboard.org/r/1390/