Fix the default changeset support with more recent versions of P4.
The new default changeset support was failing to set a default changeset
description on newer versions of P4, which we would later need to parse
in order to get the file listing. We now go through the same code path we went
through with older versions of P4 in order to generate a default description
in order to get those generated files.

Patch by Laurent Nicolas.
Reviewed at http://reviews.reviewboard.org/r/1418/