Check for GNU diff when needed.
Another common problem in RBTools is that people try to use Subversion
without having GNU diff installed. We require GNU diff to work around some
bugs in Subversion's differ.

Now, when we've determined we have a Subversion repository, we perform a check
for GNU diff. If it's not installed, we error out with an explanation. If
running on Windows, we also link to the download page for a standalone
GNU diff.

Fixes bug #769
Reviewed at http://reviews.reviewboard.org/r/1488/