Perforce: Speed up creating review request diff for a path/range diff
Patch from Gyula Faller. If you posting a review request in a Perforce workspace
from a range of changes for a range of files, and not only one changelist, it is
very slow now, since it diffs all the files in the range.  But there is speed
up: When the revision numbers equal in Perforce, the two files are identical, by
definition.