Don't use the login cookie if username/password or submit-as are provided.
If a user is giving submit-as or a username and password on the command line,
it's unlikely that they want to log in using the session stored in the cookie.
This change moves the cookie check until after we've looked at the command-line
arguments. This is based loosely on http://reviews.reviewboard.org/r/713/ but is
a different implementation.

Fixes bug 762.