Update post-review to work with the Review Board 1.5 API.
This updates post-review to use the new API in Review Board 1.5 when
available. The deprecated API won't be used in this case. However, if
the new API is not available, we fall back on the deprecated API in order
to continue supporting older installations.

This will require some fixes/enhancements to the API that are going in to
1.5.2. The API checks will be updated to check for the version once that
release has been done.