Fix the --username= option and prompt after invalid credentials.
We now support pre-setting the authorization information when --username=
is used. We have a new PresetHTTPAuthHandler class that sets the
header before the first request goes through.

We also now re-prompt for credentials if an authorization request failed.
Python's not very good at handling this sort of situation, so it required
a special ReviewBoardHTTPBasicAuthHandler class that could only retry once
and then fail, allowing our existing retry with forced login code to go
through.