Fix an authentication breakage with Python 2.6.6.
In Python 2.6.6, a self.retried variable was introduced in BasicHTTPAuth,
which conflicted with our own self.retried variable, causing odd behavior.
Basically, it prevented the code from asking the user for credentials.
This wasn't a problem in 2.6.5.

To fix it, we now have a different name for our variable. Thanks Python.