Fix problems with detecting relative paths when using Subversion.
The code for dealing with relative paths to compute the base path was
failing on some Subversion repository setups. We would bail prematurely
in a comparison and fail to post a diff.

This fixes the logic, and adds unit tests to ensure it doesn't break again.

Patch by Andrew Stitcher, with unit tests by me.
Reviewed at http://reviews.reviewboard.org/r/1957/