Check for .git/svn before calling 'git svn info'.
The call to 'git svn info' has some side-effects. It will create a .git/svn
directory, even if not using a git repository. It does this through a database
migration script. The call is generally fast, but on large repositories can
be quite slow.

We now check for the existence of the directory and whether there are any
contents before attempting to call 'git svn info'.

This actually speeds up normal git usage considerably.