Support move/delete changes in Perforce.
Perforce supports "move/delete" changes, and post-review wasn't handling this.
We now handle this in our existing "delete" handler for now.

In the future, we'll want to provide metadata telling Review Board that the
file has moved, for right now it'll be represented as a delete and an add.

Based on a patch by Gordon McGregor.