Add new "--p4-passwd" command line option
Adds the "--p4-passwd" command line option to allow a user to supply the
Perforce Ticket or clear-text password.  This is useful in cases where the
Perforce password/ticket is not available in the environment when the command
is run (such as, when this command is part of an automated script).

Reviewed at http://reviews.reviewboard.org/r/1766/