Add support for specifying the repository in .reviewboardrc
Scanning for a repository path is, by default, a good thing, but some
setups are more complicated with a variety of repository paths all
pointing to the same place. For example, SSH-backed paths, where a username
comes into play.

This allows a .reviewboardrc file to specify the repository with the
REPOSITORY variable. Along with a path, it will also support a repository ID,
or a repository name (if using Review Board 1.5.3+).

This also centralizes the config loading so it happens in one place, now that
more than one part of the code needs access to it.