Fix version detection when deciding whether to force the old API for changenums.
We were checking whether Review Board 1.5.2 was used in order to determine
whether or not to use the new API when changenums were used. We were keying
off from 1.5.2. However, I accidentally left in code that forced the
version number to 1.5.2, which guaranteed we'd never use the new API on any
Review Board version when changenums were involved.

This turned out not to be a bad thing, because we also needed the check to
cover up to 1.5.3.1, which it now does.