Add automatic parent diff support for git reviews with revision-range.
When using a revision-range, you need to specify a parent diff if the first
revision is not available in the upstream which reviewboard can see (e.g. via
gitweb).  Hence I have made this happen automatically by detecting whether the
upstream contains that revision and then creating a parent diff from the
merge-base to the first revision.

Previously it appeared that I had to do this by creating diffs manually and
uploading them to the web interface, so this improves matters somewhat for me.

Reviewed at http://reviews.reviewboard.org/r/2378/