Lines starting with '---', '+++' aren't necessarily diff control lines.
We ran into this bug with a file containing a database schema.  It was well
commented, so a lot of lines started with '-- ' followed with comment text.
Unfortunately, when these lines were deleted the resulting raw diff contained
'--- ', and as the raw diff was processed those lines would hit this code.

We were able to submit a properly-formed diff of the schema file to ReviewBoard
with this patch applied.  Without it, the diff was malformed.

Reviewed at http://reviews.reviewboard.org/r/2357/