Use --repository_url to override the git origin URL
This patch allows the user to override the current git repository's origin URL
by setting --repository_url.

We use this where I work with a custom ReviewBoard/Github:FI integration. We
have canonical repos in Github that contributors fork. Contributors work in
their fork while a script on the ReviewBoard server clones the canonical repo
and adds a remote/fetches commits for every fork so that ReviewBoard has access
to all commits.

The repository URL that our integration needs in order to use post-review is the
canonical repo, and not the fork. --repository-url seems like a reasonable entry
point to add this functionality.

In use in our organization by our engineering group.

Reviewed at http://reviews.reviewboard.org/r/2376/