Pull out code of SCM clients from postreview into separate files
This is the initial work for making RBTools an extensible set of tools with
a Python API. The various SCMClients have been pulled out into their own
files, as have many of the utility functions that exist in post-review used
by the clients.

The client design is much the same as it was, with their own options and
config data associated with the client, instead of being a global part of
post-review.

This should in the future allow for third parties to provide custom clients
without as much work, and is an important stepping stone to merging in the
new API work.

Reviewed at http://reviews.reviewboard.org/r/2399/