Fix breakages in the unit tests.
My change from a while back to add support for multiple configs ended
up breaking unit tests, and due to other test infrastructure problems,
was never caught. This boiled down to not properly resetting the config
data, not loading in the config files after writing a .reviewboardrc file,
and not defining a 'repository_url' default for OptionsStub. The unit tests
now pass.