The output of svn diff confuses ReviewBoard because renamed dir paths (which do not exist until after this commit) are probed in the repository and thus *do not exist*.
This method finds instances of this issue, and mangles the new paths
to use their old path names. Yes, this displays misleading information
in the ReviewBoard ticket, but is the a way (only way?) to make this work for
ReviewBoard+Subversion.

Function `handle_renames(diff_content)` doesn't work for the case of
directory renames because the files contained within the renamed
directory aren't considered "renamed" (and don't contain a 'Copied
From URL' svn info key).