Fix an error when posting a review inside a mercurial branch.
The _get_top_and_bottom_outgoing_revs method for the MercurialClient
was assuming all revision numbers were consecutive which is not necessarily
true at all times, especially when branches exists. This change tries to
fetch the "parent" revision instead of just subtracting one to the bottom_rev
to be reviewed, if this revision is not found it falls back to the old
behavior.

Reviewed at http://reviews.reviewboard.org/r/2487/