Rework the 'rb' command to better handle option parsing and command listing.
The 'rb' command now uses optparse to handle all the option parsing, and
lists all the commands that can be used in the description in the help text.
The hard-coded references to 'rb' and 'rb-<command>' have been moved to
consts in order to keep things standardized.

Reviewed at http://reviews.reviewboard.org/r/2549/