Compare the realm, rather than the URL, when deciding on authentication.
The authentication code was falling back on the built-in Python HTTP
Basic Auth support when the URL wasn't what we expected. This would
break post-review in the case where the URL is old and redirects, or the
server otherwise reports a different URL.

We now compare the realm when deciding, instead of the URL. This is a
better indicator and fixes this problem.