Informative error message on not being able to locate GNU diff for Perforce
Currently, if GNU diff is not in the path, post-review gives an unhandled
exception and a call stack. It scares the user, and left him without any clue
what is wrong. I found this is solved in case of SVN so I ported that solution
to Perforce too.

Fixes bug 2352
Reviewed at http://reviews.reviewboard.org/r/2685/