Fix handling of new files in perforce post-commit scenarios.
It is a erroneous asumption that an added file has no revisions yet. It is true
for pre-commit reviews, but far from true in post commit reviews. So if a file
is added it must be explicitly specified that we need its first revision.
Otherwise we will get its last revision by default.

Fixes bug 2369
Reviewed at http://reviews.reviewboard.org/r/2686/