Fix a broken link.
When running post-review against a server with no configured repositories, it
would print out a link to the documentation for adding them. Unfortunately, when
the docs got rearranged, this link got broken. This had already been fixed on
release-0.3.x but never made it to master. Trivial fix.

Fixes bug 2482.