Allow using --diff-filename=- with a valid cookie.
When using --diff-filename=-, there was a check to make sure that users wouldn't
have to type in their username and password (since stdin was used up by the
pipe), but this check didn't take into account that there might be a cookie with
the authentication information. This made it okay for automation but kind of
lame for people who actually wanted to use pipes themselves.

Fixes bug 2052.