Fix handling of revisions with deleted files for SVN.
The implementation of post-review did the wrong thing when the SVN revision had
a deleted file. Posting the diff manually worked fine, so the bug was entirely
in the post-review side.

This patch has been kicking around in our bug tracker for more than a year, in
formats which made it very hard to review. Multiple posters there vouched that
it worked.

Fixes bug 1113.