Fix issues running post-review within a git-submodule with recent git versions.
On recent versions of Git, the .git directory for a submodule was
relocated to being in the top-level project instead of in the worktree
of the submodule.

This change attempts to get the top-level directory, if we can get it,
falling back on the old behavior for older versions of Git.

Fixes bug #2526
Reviewed at http://reviews.reviewboard.org/r/2960/