Print debugging info when a subcommand fails, even when ignore_errors is true.
When a subcommand gives an error and ignore_errors is false, we die with
the error output.  But when ignore_errors is true, there's no way to see
the error output, even when running with -d.  This change fixes that but
emitting the error output in debug mode.

I also changed the log messages a bit when running a command, to make it
clear that we're running a subcommand.

Reviewed at http://reviews.reviewboard.org/r/3007/