Handle modifications inside moved/copied directory for Subversion.
This change fixes an issue with post-review when a file inside a copied
path is modified (or copied/moved). In that case, 'svn info' on the leaf
path does not show 'Copied From URL' line - it is necessary to go up to
the parent path(s) to figure out the copy-from information. Once a path
with copy-from information is found, it is not necessary to go up
further.

Reviewed at http://reviews.reviewboard.org/r/3103/