Fix python 2.4 compat and allow item retrieval from lists using a pk in the API
Fix the session cookie code to be compatable with python 2.4.

The attributes were not added to the return value of urlparse until
python 2.5, so the index numbers are now used instead.

Add support for retrieving item resources from a list using the primary key.

List resources now have a 'get_item' method which takes a single
argument specifying the pk of the item to retrieve.

Automatically add 'api/' to the end of the url when initializing the API.

Reviewed at http://reviews.reviewboard.org/r/3343/