Increase the amount of useful debug information we provide.
We now log the Python version, OS, and the current directory in the
post-review debug log. This will help reduce the amount of questions we
have to ask when collecting support data.

We also now have some helpful text saying to ignore some of the errors
shown, as they're mostly harmless.