Add the users HOME in the global config dict.
BUG: If the user has set up only one .reviewboardrc in his/hers home directory,
it will not be included in the search path for configs. It's contents is
returned from load_config_files() and assigned to user_configs in main. How ever
this variable is never used in parse_options().

FIX: Add the HOME path to the search path for global configs.

Reviewed at http://reviews.reviewboard.org/r/3171/