fix review request in svn when p4 is installed
This issue occurred when post-review in an svn repo while p4 client is
also installed. Since p4 tool is checked in prio, the svn repo will be
treated in p4 way and an error message will be prompted. It would be
better we consider svn whose metadata is stored locally in prio to p4
whose metadata is stored on server side.