Base command class and 'rb post' command.
This includes the beginnings of the new base class for rb commands,
and the rb post command.

'rb post' is meant to eventually replace post-review for the common
task of generating a diff and creating/updating a review request. It
has many of the extra options of post-review stripped out, to make
it more focused.

"rb <command>" now uses the new entry points, expecting subclasses
of ``Command``. If the command isn't found it will attempt to
execute "rb-<command>" on the system.