Add initial command help
'rbt help' can now be called to display a list of commands and usage
text. 'rbt help <command>' will output the help specific to a command
including the options it takes. Formatting and the  extensiveness of
help for each command will be improved in the future, but this is a
good start.