Fix decoding and make it consistent
A few mimetypes were special cased to change the name of the field
which the response body would be accessed through which was confusing
and inconsistent. Now mimetypes which are not JSON will have the
payload accessible through the 'data' field.

Also fixes a mimetype parsing bug where text/plain was not recognized
due to some servers providing an encoding with the mimetype.