Refactor get_root to return an RBClient as well
The get_root method has been renamed get_api and will now return the
RBClient instance along with the root resource. This allows commands
to receive direct access to the API client which can be useful if they
need to update login information after instantiating the client.

A few of the commands have also been updated to make less requests to
the Review Board server by taking advantage of the uri templates on
the root resource.