Don't block waiting for user input from svn.
Added the --non-interactive option to the "svn info" command. This was done so
it would not hang waiting for user input in certain situations like using it on
an https repository path where it can ask if you want to save the password.

Testing done:
This was discovered last week when running post-review as part of a post-commit
hook. It has been tested by being used on a daily basis by a group of
developers.

Reviewed at http://reviews.reviewboard.org/r/3793/