Add option to open a browser to rbt post
A number of options were removed when creating the post command from
the post-review code. This change adds the '[-o | --open]' option back
which can be used to open a browser to the review request after it has
been created.

Also, a description was added to the post command, and two unused
imports were removed.