Fix treating an SVN repo as Perforce by mistake.
This changes the order in which we check for repositories. Perforce was
coming before Subversion, but Perforce would give us a false-positive
due to the fact that it says we have a valid repository so long as
P4PORT is set to something real.

This should be fixed in a more bullet-proof way later, but for now, this
at least makes it possible to use both Subversion and Perforce
individually.

Reviewed at http://reviews.reviewboard.org/r/3525/