Add a method for retreiving resources from a path
The RBClient now has a 'get_path' method which will accept a path and
retreive the API resource at that path. For example a path of
'/review-requests/123' would retreive the resource at
'http://example.com/api/review-requests/123'

A 'ParseError' Exception has also been added for commands. If there is
an error parsing the command arguments or options, the command should
raise a ParseError.