Introduce the api-get command.
The new rbt api-get command takes a path to an api resource, requests
the resource, and prints it as JSON. The path can either be rooted
at the API level (e.g. '/review-requests/123/') or be a full URL.

Query arguments may also be specified for the request. Each query
argument appears after '--' on the command line and takes the form
'--<query-arg>=<value>'. For example,
'rbt api-get /review-requests/ -- --counts-only=1' would result in
a request to 'http://example.com/api/review-requests/?counts-only=1'.

To support this new command, a new 'get_url' is now required for
transports. This change provides the required method for the
SyncTransport.

Reviewed at http://reviews.reviewboard.org/r/4195/