Fix rbt post and diff for certain SCM clients.
post-review provided the command line arguments to SCM clients as a
list. Each rbt command is provided the arguments as a tuple, which was
being passed on to the SCM clients. Since some of these clients
attempt to concatenate the tuple onto the end of a list, errors were
occuring.

We now make the command line arguments into a list immediately in post
and diff. This should fix bugs when using rbt with both svn and bazzar
(possibly other clients as well).

Reviewed at http://reviews.reviewboard.org/r/4111/