Fix Plastic's usage of self._options.
When the plastic client is creating a diff for a branch, it tries to set the
'branch' field in the options object so that it will get set properly in the
review request. This broke when we refactored the command-line options to be
able to have different options for each rbt command.

Addresses bug 3044
Reviewed at https://reviews.reviewboard.org/r/4395/