Send the base commit ID for a change to the server when using Git.
When uploading a change to a new enough Review Board server, the base
commit ID is sent along in the request. This ID/revision represents the
upstream commit that the diff (or its parent diff) applies to. It can be
used by the server to communicate with hosting services that otherwise
do not work solely with the content found in the diffs.