Makes --revision-range behave like documented for one revision in git
The documentation states that "If you only need to post a single revision,
you can type: $ rbt post --revision-range=REVISION", however that was not the
case. This changes makes rbt post --revision-range=REVISION act like you had
given it --revision-range=REVISION^:REVISION which is how the documentation
makes it sound like it should behave.

NOTE: this one changes the behavior for git. Same behavior is apparently
present for other tools.