Fix git parent diff support in diff_between_revisions
The calling contract stipulates that the parent diff should be in a
dictionary entry called 'parent_diff', not 'parent_diff_lines'.

I also snuck in a change to the git unit tests: if they don't create a
dummy .gitconfig in the new HOME, "git commit" dies later, triggering all
sorts of unhappiness in nose.