Enhance rbt post to automatically determine the existing review request
``rbt -u`` automatically determines the existing review request to update. It
suggests a pending review request by the current user in the current
repository, based on the request summary and description.

If the summary and description exactly match those of an existing request, the
request is immediately updated. Otherwise, the user is prompted to confirm a
suggestion from a ranked list of potential matches.

Testing Done:
- Tested with Git, Mercurial, and Bazaar using, with and without 'revision
  range'.
- Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/4699/