Un-escape markdown text for 'rbt patch -c'
The patching step was including escaped characters in the description and
testing done, which was annoying to look at, and made it so in the case where
the first line of the description was the same as the summary except for
escaping, we wouldn't collapse them together.

This change adds some code to do the same as markdown_unscape() from
reviewboard.reviews.markdown_utils.

Testing done:
Ran `rbt patch -c` for a few changes on reviews.reviewboard.org that had
escaped descriptions and testing done. Saw the expected results.

Reviewed at https://reviews.reviewboard.org/r/5014/