Fix the authentication prompt appearing multiple times.
The recent change to reorganize how login works (9ab0cac0) actually
broke authentication a bit. RBTools would always prompt at least twice
for a password. This was due to the logic being incomplete and not
actually making use of the new parameters passed in.

This completes that change by adding those parameters instead of trying
for options.username and options.password.