Change 'check_install' to take an list of parameters instead of a string.
The 'check_install' method is used to check to see if various command-line
tools we use are present. This method used to take a string and split it on
spaces to get a list of parameters. This seemed an ugly and error-prone way of
breaking things up, so I've changed it to take in a list of parameters instead,
which are passed directly into Popen.

Testing done:
* Ran unit tests.
* Smoke tested rbt.

Reviewed at https://reviews.reviewboard.org/r/5095/