Speed up rbt post -u.
rbt post -u, in an effort to match draft summaries, was fetching every
opened draft for a user individually and pulling info from them. This
was slow, particularly with lots of pending review requests.

Now it uses the field expansion support in the API to fetch drafts in
one go, reducing the number of calls significantly.