Fix the affinity of the "pdiff_required" value.
When using --revision-range with git, we determine whether or not we need to
create a parent diff by checking if the upstream branch contains the source
revision. The result of this check went into a variable called
"pdiff_required", but the value of this variable was truthy if a parent diff is
*not* required. I've negated the assignment expression to flip this back to
mean what it says.

Testing done:
Posted a review request with --revision-range that uses a parent diff.

Reviewed at https://reviews.reviewboard.org/r/5102/