Improve the RBTools configuration docs.
This makes a number of changes to the configuration docs. First, it
fixes the location of the .reviewboardrc docs (which was nested under
Subversion, oddly).

It then expands on the .reviewboardrc docs in a few ways. It recommends
.reviewboardrc over repo-specific configuration (properties and such).
It points to 'rbt setup-repo' and rbt post's default options.

It then covers some of the common options and how to use them, like
TRACKING_BRANCH, BRANCH, and ENABLE_PROXY. These are generally good ones
people may want to set, so we want them to see these right away.

Finally, it rewrites the User Configuration section to talk more about
user-facing options, like ENABLE_PROXY (again) and OPEN_BROWSER, and
removes information about TREES and custom URLs and Python code. People
can always create a .reviewboardrc in the repo directory anyway, and
it's best to not add confusion. This will also hopefully help us move
away from supporting TREES and such down the road.